# enable -fno-math-errno as we never check errno